refactor: improve plugin architecture with modular design and enhanced tools support
- Reorganize codebase into modular structure with separate concerns
- Split functionality into dedicated modules: models/, utils/, providers/
- Add comprehensive tools support for TEXT_SMALL and TEXT_LARGE models
- Implement proper tool execution handling and empty response fallbacks
- Add structured response parsing for image descriptions
- Enhance JSON repair functionality with better error recovery
- Update core dependency to latest version (^1.2.5)
- Improve code maintainability and readability
- Add proper TypeScript types throughout the codebase
- Clean up debug logging and optimize performance
- Support tool choice parameters and multi-step tool execution
