+++ package.json
@@ -24,7 +24,7 @@
   "dependencies": {
     "@ai-sdk/openai": "^1.3.22",
     "@ai-sdk/ui-utils": "1.2.11",
-    "@elizaos/core": "^1.0.0",
+    "@elizaos/core": "^1.2.5",
     "@openrouter/ai-sdk-provider": "^0.4.5",
     "ai": "^4.3.15",
     "undici": "^7.9.0"
