+++ src/models/index.ts
@@ -1,0 +1,3 @@
+export { generateTextWithModel } from './text';
+export { generateObjectByModelType } from './object';
+export { generateImageDescription } from './image';
