+++ src/providers/index.ts
@@ -1,0 +1,1 @@
+export { createOpenRouterProvider } from './openrouter';
